+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [8.56.1] - 2025-08-29
+
+### Fixed
+- The GenericRecordDataDeserializer is now only available without signature check
+- Better logging when signature verification fails
+
 ## [8.56.0] - 2025-08-26
 
 ### Changed